--- src/components/grid/index.ts
@@ -1,1 +1,0 @@
-export * from "./Grid";
