+++ src/components/index.ts
@@ -1,3 +1,2 @@
-export * from "./grid";
 export * from "./legend";
 export * from "./sequencer";
